# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'UISS API documentation'
  description: 'API documentation for basic CRUD operations for specified requests. All API endpoints are implemented in JSON Schema. The data transaction must be HTTP requests sent via specified url'
  version: 1.0.0
servers:
  -
    url: 'https://api.uiss.patrickmamsery.co.tz'
paths:
...
  /api/events:
    get:
      summary: 'Get all events'
      description: ''
      parameters:
        -
          in: query
          name: page
          description: 'The page number to retrieve.'
          example: 1
          required: false
          schema:
            type: integer
            description: 'The page number to retrieve.'
            example: 1
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data: []
                  meta:
                    current_page: 1
                    from: null
                    last_page: 1
                    path: 'https://api.uiss.patrickmamsery.co.tz/api/events'
                    per_page: 15
                    to: null
                    total: 0
                  status: success
                  message: 'Resource retrieved successfully'
                  statusCode: 200
                properties:
                  data:
                    type: array
                    example: []
                  meta:
                    type: object
                    example:
                      current_page: 1
                      from: null
                      last_page: 1
                      path: 'https://api.uiss.patrickmamsery.co.tz/api/events'
                      per_page: 15
                      to: null
                      total: 0
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: 'Resource retrieved successfully'
                  statusCode:
                    type: integer
                    example: 200
      tags:
        - 'Event management'
      security: []
    post:
      summary: 'Create a new event'
      description: 'This endpoint creates a new event.'
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer Token'
          schema:
            type: string
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses: {  }
      tags:
        - 'Event management'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'The name of the event.'
                  example: 'Event 1'
                description:
                  type: string
                  description: 'The description of the event.'
                  example: 'This is an event'
                venue:
                  type: string
                  description: 'The venue of the event.'
                  example: 'Dar es Salaam'
                image:
                  type: string
                  description: 'The image of the event.'
                  example: image.jpg
                startDate:
                  type: string
                  description: 'The start date of the event.'
                  example: '2021-01-01'
                endDate:
                  type: string
                  description: 'The end date of the event.'
                  example: '2021-01-01'
                hosts:
                  type: string
                  description: 'The hosts of the event separated by commas and single whitespaces.'
                  example: 'John Doe, Jane Doe'
              required:
                - name
                - description
                - venue
                - startDate
                - endDate
                - hosts
  '/api/events/{id}':
    get:
      summary: 'Get a single event'
      description: ''
      parameters:
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: error
                  message: 'Failed to retrieve resource'
                  statusCode: 404
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: 'Failed to retrieve resource'
                  statusCode:
                    type: integer
                    example: 404
      tags:
        - 'Event management'
      security: []
    put:
      summary: 'Update an event'
      description: 'This endpoint updates an event.'
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer Token'
          schema:
            type: string
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses: {  }
      tags:
        - 'Event management'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'The name of the event.'
                  example: 'Event 1'
                description:
                  type: string
                  description: 'The description of the event.'
                  example: 'This is an event'
                venue:
                  type: string
                  description: 'The venue of the event.'
                  example: 'Dar es Salaam'
                image:
                  type: string
                  description: 'The image of the event.'
                  example: image.jpg
                startDate:
                  type: string
                  description: 'The start date of the event.'
                  example: '2021-01-01'
                endDate:
                  type: string
                  description: 'The end date of the event.'
                  example: '2021-01-01'
                hosts:
                  type: string
                  description: 'The hosts of the event separated by commas and single whitespaces.'
                  example: 'John Doe, Jane Doe'
    delete:
      summary: 'Delete an event'
      description: "This endpoint deletes an event and all it's relations completely.\n\n<aside class=\"notice\"> <strong>NOTE:</strong> This action is irreversible </aside>"
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer Token'
          schema:
            type: string
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses: {  }
      tags:
        - 'Event management'
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the event.'
        example: 9
        required: true
        schema:
          type: integer
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: ''
security:
  -
    default: []